# salon-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salon API",
        "description": "API za upravljanje frizerskim salonom",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/schedules": {
            "get": {
                "tags": [
                    "Rasporedi"
                ],
                "summary": "Lista svih rasporeda",
                "operationId": "f21805e74c0fcc7937d0464ff67212bd",
                "responses": {
                    "200": {
                        "description": "Lista rasporeda"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rasporedi"
                ],
                "summary": "Kreiranje ili ažuriranje rasporeda",
                "operationId": "ec2ea6666bdb71e07d0fe1bcaf240511",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "day_of_week",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "day_of_week": {
                                        "type": "string",
                                        "example": "Monday"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "09:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "17:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Raspored ažuriran"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/schedules/{id}": {
            "put": {
                "tags": [
                    "Rasporedi"
                ],
                "summary": "Ažuriranje rasporeda",
                "operationId": "a1aa67236cf075c90a0e5feee2e76629",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start_time": {
                                        "type": "string",
                                        "example": "09:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "17:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Raspored ažuriran"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rasporedi"
                ],
                "summary": "Brisanje rasporeda",
                "operationId": "7e08cf8a05dcade2a47627cf994d9021",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Raspored obrisan"
                    },
                    "404": {
                        "description": "Nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/my-schedule": {
            "get": {
                "tags": [
                    "Rasporedi"
                ],
                "summary": "Moj raspored",
                "operationId": "f11a2e31d77322f0a368c9605306b1bc",
                "responses": {
                    "200": {
                        "description": "Raspored trenutnog frizera"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/services": {
            "get": {
                "tags": [
                    "Usluge"
                ],
                "summary": "Lista svih usluga",
                "operationId": "97de511018def32ca0ccbc800756c538",
                "responses": {
                    "200": {
                        "description": "Lista usluga"
                    }
                }
            },
            "post": {
                "tags": [
                    "Usluge"
                ],
                "summary": "Kreiranje nove usluge",
                "operationId": "0634f2b3ac65bd2815f4b6459f0db97c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "duration_minutes",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Šišanje"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Osnovno šišanje"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 1500
                                    },
                                    "duration_minutes": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usluga kreirana"
                    },
                    "422": {
                        "description": "Greška validacije"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "Usluge"
                ],
                "summary": "Detalji usluge",
                "operationId": "dce5b306a141bdec91eff8a15534d7d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalji usluge"
                    },
                    "404": {
                        "description": "Usluga nije pronađena"
                    }
                }
            }
        },
        "/api/services/{service}": {
            "put": {
                "tags": [
                    "Usluge"
                ],
                "summary": "Ažuriranje usluge",
                "operationId": "3a9545bd0acf658bb3b26cbc87029665",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Šišanje"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 1800
                                    },
                                    "duration_minutes": {
                                        "type": "integer",
                                        "example": 45
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usluga ažurirana"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usluge"
                ],
                "summary": "Brisanje usluge",
                "operationId": "a700b2ae0c8b0e6ccf4812c2bba02e97",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usluga obrisana"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}